Changed default type of SSH key
This is needed because the python library that connects to SSH
(paramiko, used by AiiDA) does not recognize the new OpenSSH format
of the key that is now the default in recent linux/mac installations.
We specify then by hand the PEM format and a length of a key, that
are the ones used in earlier linux/mac installations, so that
paramiko will work with this key.

See also https://github.com/aiidateam/aiida-core/pull/3281